move interface to pkg, where used
